Small adjustments

+++ src/main/kotlin/invoice/management/system/repositories/PurchaseInvoiceRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PurchaseInvoiceRepository : JpaRepository<PurchaseInvoice, Long>
+interface PurchaseInvoiceRepository : JpaRepository<PurchaseInvoice, Int>
